Refaktoroi konfiguraation lataaminen ja tyypitys
