+++ valpas-web/src/index.tsx
@@ -1,33 +1,29 @@
 import * as E from "fp-ts/Either"
+import { pipe } from "fp-ts/lib/function"
 import React from "react"
 import ReactDOM from "react-dom"
 import "regenerator-runtime/runtime"
-import { fetchWindowProperties } from "./api/api"
+import { fetchAppConfiguration } from "./api/api"
+import { AppConfiguration } from "./state/apitypes/appConfiguration"
 import { enableFeature, Feature } from "./state/featureFlags"
 import "./style/index.less"
 import { ValpasApp } from "./views/ValpasApp"
 
 declare global {
-  interface Window {
-    environment?: string
+  interface Window extends AppConfiguration {
     virkailija_raamit_set_to_load?: boolean
-    opintopolkuVirkailijaUrl?: string
     enableFeature?: (feature: Feature) => void
   }
 }
 
-async function loadWindowProperties() {
-  const response = await fetchWindowProperties()
-  if (E.isRight(response)) {
-    const props = response.right.data
-    window.valpasLocalizationMap = props.valpasLocalizationMap
-    window.environment = props.environment
-    window.opintopolkuVirkailijaUrl = props.opintopolkuVirkailijaUrl
-  }
-  if (E.isLeft(response)) {
-    console.error("Konfiguraation haku epäonnistui:", response.left)
-  }
-}
+const loadWindowProperties = async (): Promise<void> =>
+  pipe(
+    await fetchAppConfiguration(),
+    E.fold(
+      (error) => console.error("Konfiguraation haku epäonnistui:", error),
+      (props) => Object.assign(window, props.data)
+    )
+  )
 
 async function main() {
   await loadWindowProperties()
