+++ valpas-web/src/utils/tableDataFormatters/hakemuksentila.tsx
@@ -5,7 +5,8 @@ import React from "react"
 import { Link } from "react-router-dom"
 import { ExternalLink } from "../../components/navigation/ExternalLink"
 import { Value } from "../../components/tables/DataTable"
-import { getLocalizedMaybe, t, Translation } from "../../i18n/i18n"
+import { getLocalizedMaybe, t } from "../../i18n/i18n"
+import { Translation } from "../../state/apitypes/appConfiguration"
 import { HakuSuppeatTiedot } from "../../state/apitypes/haku"
 import { OppijaHakutilanteillaSuppeatTiedot } from "../../state/apitypes/oppija"
 import { Oid } from "../../state/common"
